Update BAAD-TicTacToe.cpp
"Needs to loop for full game and check for win and draw, then give option to restart and quit."

Took these notes into consideration when redoing the code. Looped the full game, made the code so that it would check for wins and draws, and finally it gives the option to either restart or quit.